requirements.txt
requests>=2.31,<3
feedparser>=6.0,<7
gspread>=6.0,<7
google-auth>=2.28,<3
PyYAML>=6.0,<7
